Refactored main structure of the project.
