+++ .travis.yml
@@ -1,0 +1,6 @@
+language: python
+
+python:
+  - "3.6"
+
+script: python load_tests.py
