+++ requirements.txt
@@ -1,6 +1,11 @@
-# Flask
+click==6.7
 Flask==1.0.2
-sqlalchemy==1.2.7
-alembic==?
-
-# PostgreSQL
+Flask-Script==2.0.6
+Flask-SQLAlchemy==2.3.2
+itsdangerous==0.24
+Jinja2==2.10
+MarkupSafe==1.0
+psycopg2==2.7.5
+redis==2.10.6
+SQLAlchemy==1.2.7
+Werkzeug==0.14.1
